wip: Elasticsearch 2.0
* Change all `float32` to `float64`.
* Update README.md to introduce Elastic 3.0 and document supported
  features and examples.
* Prepare a first draft of the changes that Elastic 3.0 introduces.

+++ search.go
@@ -432,7 +432,7 @@ type SearchSuggestion struct {
 // See http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-suggesters.html.
 type SearchSuggestionOption struct {
 	Text    string      `json:"text"`
-	Score   float32     `json:"score"`
+	Score   float64     `json:"score"`
 	Freq    int         `json:"freq"`
 	Payload interface{} `json:"payload"`
 }
